Ported most infobox functions to the main program since INSPQ stopped publishing on website

+++ main.py
@@ -2,8 +2,45 @@ import requests
 import re
 from os import path
 import os
+from datetime import datetime, timedelta
 
 fileDir = os.path.dirname(os.path.realpath('__file__'))
+
+# If running program on a weekend, set date to the last Friday
+if datetime.now().weekday() == 5:
+    todaysDate = datetime.now() - timedelta(days=1)
+elif datetime.now().weekday() == 6:
+    todaysDate = datetime.now() - timedelta(days=2)
+else:  # Otherwise set to today's date
+    todaysDate = datetime.now()
+
+offset = (datetime.now().weekday() - 1) % 7
+mtlDate = datetime.now() - timedelta(days=offset)
+mtlDate = f'{mtlDate:%B} {mtlDate.day}, {mtlDate.year}'  # Format date
+todaysDate = f'{todaysDate:%B} {todaysDate.day}, {todaysDate.year}'
+
+# Figure out how better to do this later
+dates = []
+cases = []
+activeCases = []
+dateDeaths = []
+newCases = []
+deaths = []
+hospitalizations = []
+deathsTotal = []
+percentage1st = []
+dateVaccination = []
+dailyDoses1st = []
+dailyDoses2nd = []
+dailyDoses3rd = []
+dailyDosesTotal = []
+
+totalDoses1st = []
+totalDoses2nd = []
+totalDoses3rd = []
+totalDosesTotal = []
+
+percentage2nd = []
 
 
 # Helper function to compute the average of a list
@@ -11,10 +48,75 @@ def computeAverage(lst):
     return sum(lst) / len(lst)
 
 
+def smallDate(date):
+    return "<small>(" + date + ")</small>"
+
+
+def createAttribute(name, val):
+    attr = {
+        "name": name,
+        "value": val
+    }
+
+    return attr
+
+
+def openFile(filePath):
+    if path.exists(filePath):  # Check if path already exists
+        fileName = open(filePath, "w", encoding='utf-8')
+    else:  # Create file if it doesn't exist
+        fileName = open(filePath, "x", encoding='utf-8')
+    return fileName
+
+
 def downloadCSV(url, fileName):
-    # Download and save file
-    r = requests.get(url, allow_redirects=True)
-    open(fileName, 'wb').write(r.content)
+    try:
+        # Download and save file
+        r = requests.get(url, allow_redirects=True)
+
+        if os.path.isfile(fileName):
+            open(fileName, 'wb').write(r.content)
+    except PermissionError:
+        print(f"Error: you currently have {fileName} open. Please close it and try running the program again.")
+        exit()
+
+
+def readCSV(csv):
+    csv.readline()  # Skip first line
+
+    casesFlag = False
+    deathsFlag = False
+    hospitalizationsFlag = False
+
+    # Parse file
+    for x in csv:
+        line = re.split(',', x)
+        if (not line[0]) or (len(line) <= 1):
+            break
+
+        if (not casesFlag) and (line[0] == '2020-02-24'):
+            casesFlag = True
+
+        if (not deathsFlag) and (line[0] == '2020-03-15'):
+            deathsFlag = True
+
+        if (not hospitalizationsFlag) and (line[0] == '2020-04-10'):
+            hospitalizationsFlag = True
+
+        # Gather data and add to arrays
+        if casesFlag and (line[2] == "RSS99"):
+            dates.append(line[0])  # Add dates into array
+            newCases.append(int(line[11]))  # Add new cases into array
+            cases.append(int(line[6]))
+            activeCases.append(int(line[12]))
+
+        if hospitalizationsFlag and (line[2] == "RSS99") and line[44] != '.':
+            hospitalizations.append(int(line[44]))
+
+        if deathsFlag and (line[1] == 'Sexe') and (line[2] == 'TOT'):
+            dateDeaths.append(line[0])  # Add dates into array
+            deaths.append(line[25])  # Add deaths into array
+            deathsTotal.append(int(line[18]))
 
 
 def montreal():
@@ -80,7 +182,14 @@ def quebec():
     downloadCSV('https://www.inspq.qc.ca/sites/default/files/covid/donnees/covid19-hist.csv', "covid19-hist.csv")
     quebecCasesCSV = open("covid19-hist.csv", "r")
 
-    quebecCasesCSV.readline()  # Skip first line
+    readCSV(quebecCasesCSV)
+
+    quebecGraphs(quebecCasesCSV)
+    quebecInfobox(quebecCasesCSV)
+
+
+def quebecGraphs(csv):
+    movingAverage = []
 
     # File names
     QuebecNewCases = "Files_Quebec/QuebecNewCases.txt"
@@ -132,36 +241,6 @@ def quebec():
                              "|xType = date\n",
                              "|x = "])
 
-    date = []
-    dateDeaths = []
-    newCases = []
-    movingAverage = []
-    deaths = []
-
-    casesFlag = False
-    deathsFlag = False
-
-    # Parse file
-    for x in quebecCasesCSV:
-        line = re.split(',', x)
-        if (not line[0]) or (len(line) <= 1):
-            break
-
-        if (not casesFlag) and (line[0] == '2020-02-24'):
-            casesFlag = True
-
-        if (not deathsFlag) and (line[0] == '2020-03-15'):
-            deathsFlag = True
-
-        # Gather data and add to arrays
-        if casesFlag and (line[1] == 'Sexe') and (line[2] == 'TOT'):
-            date.append(line[0])  # Add dates into array
-            newCases.append(int(line[11]))  # Add new cases into array
-
-        if deathsFlag and (line[1] == 'Sexe') and (line[2] == 'TOT'):
-            dateDeaths.append(line[0])  # Add dates into array
-            deaths.append(line[25])  # Add deaths into array
-
     # Calculate 7-day moving average
     i = 0
     j = 7
@@ -175,7 +254,7 @@ def quebec():
     # Write to file
     quebecCases = open(QuebecNewCases, "a")
 
-    for x in date:
+    for x in dates:
         quebecCases.write(x + ",")
 
     quebecCases.writelines(["\n|y1Title=Daily new cases\n", "|y1="])
@@ -202,10 +281,64 @@ def quebec():
     quebecDeaths.close()
 
 
-def vaccinations():
+def quebecInfobox(csv):
+    refs = ["<ref name=\"auto6\">{{Cite "
+            "web|url=https://www.inspq.qc.ca/covid-19/donnees|title=Données "
+            "COVID-19 au Québec|website=INSPQ}}</ref>",
+            "<ref name=\"auto6\"/>",
+            "<ref name=\"inspqVacc\">{{cite web |title=Données de "
+            "vaccination contre la COVID-19 au Québec "
+            "|url=https://www.inspq.qc.ca/covid-19/donnees/vaccination "
+            "|website=INSPQ |publisher=Gouvernement "
+            "|access-date=2021-03-19|language=fr}}</ref>",
+            "<ref name=\"vacc\">{{cite web |title=DATA COVID-19 VACCINATION IN MONTRÉAL "
+            "|url=https://santemontreal.qc.ca/en/public/coronavirus-covid-19/vaccination/data-vaccination/ "
+            "|website=Santé Montréal |publisher=Gouvernement du Québec}}</ref>",
+            "<ref name=\"vacc\"/>"
+            ]
+
+    QCPathName = "infoboxes/QuebecInfobox.txt"
+    QCFile = openFile(QCPathName)
+
+    # Format infobox
+
+    efn = "{{efn|This figure may not represent the current epidemiological situation — the Quebec government " \
+          "restricted PCR COVID-19 tests to certain vulnerable groups on January 4, 2022.}} "
+
+    currentDate = datetime.strptime(dates[-1], '%Y-%m-%d')
+    currentDate = datetime.strftime(currentDate, '%B %d, %Y')
+
+    date = createAttribute("date", currentDate)
+    confirmedCases = createAttribute("confirmed_cases", f'{cases[-1]:,}{refs[0]}')
+    activeCase = createAttribute("active_cases", f'{activeCases[-1]:,}{efn}')
+    death = createAttribute("deaths", f'{deathsTotal[-1]:,}{refs[1]}')
+    fatalityRate = createAttribute("fatality_rate",
+                                   "{{Percentage|" + str(deathsTotal[-1]) + "|" + str(cases[-1]) + "|2}}")
+    hospitalization = createAttribute("hospitalized_cases", f'{hospitalizations[-1]:,}')
+    vax = createAttribute("vaccinations",
+                          "\n*'''" + f'{float(percentage1st[-1]):.1f}' + "%'''   vaccinated with at least one dose " +
+                          refs[2])
+
+    infobox = [date, confirmedCases, activeCase, death, fatalityRate, hospitalization, vax]
+
+    text = ""
+    for attribute in infobox:
+        text += f"| {attribute['name']}          = {attribute['value']}\n"
+
+    QCFile.write(text)
+    QCFile.close()
+
+
+def vaccination():
     downloadCSV('https://www.inspq.qc.ca/sites/default/files/covid/donnees/vaccination.csv', "vaccination.csv")
     vaccinationsCSV = open("vaccination.csv", "r")
 
+    vaccinationGraphs(vaccinationsCSV)
+    vaccinationInfobox()
+    vaccinationPiechart()
+
+
+def vaccinationGraphs(vaccinationsCSV):
     vaccinationsCSV.readline()  # Skip first line
 
     # File names
@@ -287,26 +420,10 @@ def vaccinations():
                            "| legend=Legend\n",
                            "| y1Title=1st dose\n",
                            "| y2Title=2nd dose\n",
-     #                     "| y3Title=3rd dose\n",
+                           #                     "| y3Title=3rd dose\n",
                            "| yAxisTitle =  Percentage of the population vaccinated (%)\n",
                            "| x="
                            ])
-
-    date = []
-
-    dailyDoses1st = []
-    dailyDoses2nd = []
-    dailyDoses3rd = []
-    dailyDosesTotal = []
-
-    totalDoses1st = []
-    totalDoses2nd = []
-    totalDoses3rd = []
-    totalDosesTotal = []
-
-    percentage1st = []
-    percentage2nd = []
-    # percentage3rd = []
 
     # Parse file
     for x in vaccinationsCSV:
@@ -316,24 +433,24 @@ def vaccinations():
 
         # Gather data and add to arrays
         if line[2] == 'RSS99':
-            date.append(line[0])  # Add dates into array
+            dateVaccination.append(line[0])  # Add dates into array
 
             dailyDoses1st.append(line[4])
             dailyDoses2nd.append(line[5])
             dailyDoses3rd.append(line[6])
-            dailyDosesTotal.append(line[10])
+            dailyDosesTotal.append(line[12])
 
-            totalDoses1st.append(line[7])
-            totalDoses2nd.append(line[8])
-            totalDoses3rd.append(line[9])
-            totalDosesTotal.append(line[11])
+            totalDoses1st.append(line[8])
+            totalDoses2nd.append(line[9])
+            totalDoses3rd.append(line[10])
+            totalDosesTotal.append(line[13])
 
-            percentage1st.append(line[12])
-            percentage2nd.append(line[13])
+            percentage1st.append(line[14])
+            percentage2nd.append(line[15])
 
     dailyDoses = open(DailyDoses, "a")
 
-    for x in date:
+    for x in dateVaccination:
         dailyDoses.write(x + ",")
 
     dailyDoses.writelines(["\n|y1="])
@@ -361,7 +478,7 @@ def vaccinations():
 
     totalDoses = open(TotalDoses, "a")
 
-    for x in date:
+    for x in dateVaccination:
         totalDoses.write(x + ",")
 
     totalDoses.writelines(["\n|y1="])
@@ -387,7 +504,7 @@ def vaccinations():
 
     percentage = open(PercentageVaccinated, "a")
 
-    for x in date:
+    for x in dateVaccination:
         percentage.write(x + ",")
 
     percentage.writelines(["\n|y1="])
@@ -404,8 +521,85 @@ def vaccinations():
     percentage.close()
 
 
+def vaccinationInfobox():
+    refs = ["<ref name=\"inspqVacc\">{{cite web |title=Données de " \
+            "vaccination contre la COVID-19 au Québec " \
+            "|url=https://www.inspq.qc.ca/covid-19/donnees/vaccination |website=INSPQ " \
+            "|publisher=Gouvernement |access-date=2021-03-19|language=fr}}</ref>",
+            "<ref name=\"inspqVacc\"/>"
+            ]
+
+    vaxPathName = "infoboxes/vaccination.txt"
+    vaxFile = openFile(vaxPathName)
+
+    currentDate = datetime.strptime(dateVaccination[-1], '%Y-%m-%d')
+    currentDate = datetime.strftime(currentDate, '%B %d, %Y')
+    currentDate = smallDate(currentDate)
+
+    totalAdministered = f"'''{int(totalDosesTotal[-1]):,}''' doses administered {currentDate}{refs[0]}<br>"
+    total2Administered = f"'''{int(totalDoses2nd[-1]):,}''' second doses administered {currentDate}{refs[1]}"
+    participants = createAttribute("participants", totalAdministered + total2Administered)
+
+    outcome1 = f"'''{float(percentage1st[-1]):.1f}%''' of the population has received at least one dose of a vaccine {currentDate}{refs[1]}"
+    outcome = createAttribute("outcome", outcome1)
+
+    infobox = [participants, outcome]
+
+    text = ""
+    for attribute in infobox:
+        text += f"| {attribute['name']}          = {attribute['value']}\n"
+
+    vaxFile.write(text)
+    vaxFile.close()
+
+
+def vaccinationPiechart():
+    quebecPopulation = 8585523  # Stats Can estimate Q2 2021
+    currentDate = datetime.strptime(dateVaccination[-1], '%Y-%m-%d')
+    currentDate = datetime.strftime(currentDate, '%B %d, %Y')
+
+    vaxPathName = "infoboxes/PiechartQuebecVaccination.txt"
+    vaxFile = openFile(vaxPathName)
+
+    # Calculations for piechart
+
+    unvaccinated = int(((100 - float(percentage1st[-1])) / 100) * quebecPopulation)
+    unvaccinatedPercentage = round((unvaccinated / quebecPopulation) * 100, 1)
+
+    oneDose = int(((float(percentage1st[-1]) / 100) * quebecPopulation) - int(totalDoses2nd[-1]))
+    oneDosePercentage = round((oneDose / quebecPopulation) * 100, 1)
+
+    secondDosePercentage = round(float(percentage1st[-1]) - oneDosePercentage, 1)
+
+    caption = createAttribute("caption", f"Total number of people receiving vaccinations in Quebec as of {currentDate}")
+    ref = createAttribute("ref", "https://www.inspq.qc.ca/covid-19/donnees/vaccination")
+
+    label1 = createAttribute("label1",
+                             f"Unvaccinated population: ~{unvaccinated:,} people <!-- Quebec population estimate as of Q2 2021: 8,585,523 -->")
+    value1 = createAttribute("value1", unvaccinatedPercentage)
+    color1 = createAttribute("color1", "#BFBFBF")
+
+    label2 = createAttribute("label2", f" Population who has received only one dose of a vaccine: {oneDose:,} people")
+    value2 = createAttribute("value2", oneDosePercentage)
+    color2 = createAttribute("color2", "#42f5da")
+
+    label3 = createAttribute("label3",
+                             f" Population who has been fully vaccinated (both doses): {int(totalDoses2nd[-1]):,} people")
+    value3 = createAttribute("value3", secondDosePercentage)
+    color3 = createAttribute("color3", "#008")
+
+    infobox = [caption, ref, label1, value1, color1, label2, value2, color2, label3, value3, color3]
+
+    text = ""
+    for attribute in infobox:
+        text += f"| {attribute['name']}              = {attribute['value']}\n"
+
+    vaxFile.write(text)
+    vaxFile.close()
+
+
 if __name__ == "__main__":
     # Generate all files
+    vaccination()
     montreal()
     quebec()
-    vaccinations()
